feat: add codehash-based authorization for RecorderCombinator, input validation hardening
- BaseRecorder: add AUTHORIZED_CODEHASH immutable for codehash-based caller
  verification (allows RecorderCombinator to call record() on behalf of operators)
- RecorderCombinator: add OnlyOperator check so only the actual operator can
  invoke the combinator
- EscrowPeriodConditionFactory: include authorizedCodehash in deploy/lookup key
- FreezePolicy: add zero-address checks for freeze/unfreeze conditions
- StaticFeeCalculator/Factory: add FeeTooHigh validation (feeBps <= 10000)
- PaymentOperator: remove unnecessary uint120 cast in RefundExecuted emit

+++ src/operator/payment/PaymentOperator.sol
@@ -347,7 +347,7 @@ contract PaymentOperator is ReentrancyGuardTransient, PaymentOperatorAccess {
 
         // ============ EFFECTS ============
         // Emit event before external calls (CEI pattern)
-        emit RefundExecuted(paymentInfo, paymentInfo.payer, uint120(amount));
+        emit RefundExecuted(paymentInfo, paymentInfo.payer, amount);
 
         // ============ INTERACTIONS ============
         // Forward to escrow's refund - token collector enforces permission
